Added play and quit button functionality
Clicking the play button now brings user to the Minimal_Default map. The level it brings you to can be changed by going to Content>HomeScreen>MainMenu, clicking "Graph" in the top right, and changing the name in the "Open Level (by Name)" widget which should be somewhere in the center of the screen.

I'm still trying to figure out a way to temporarily disable the Main Menu since it does take a while for the level to load after clicking the New Game button which might get annoying when debugging.